backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
paramiko==3.2.0
python-dotenv==1.0.0
PyJWT==2.10.1
